requirements.txt
fastapi
uvicorn
openai
ray
requests
pydantic
pytest
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
